updated viewLessonMaterials as a component
