+++ neuro-app/src/components/LessonMaterials/Body.module.css
@@ -1,0 +1,37 @@
+.body {
+    min-height: 100%;
+    background-size: cover;
+    position: fixed;
+    width: 100%;
+    height: auto;
+}
+
+.body--text {
+    background-repeat: no-repeat;
+    background-size: cover;
+    height: 100%;
+    position:relative;
+    
+    margin-top: 3%;
+
+    padding-top: 2%;
+    padding-bottom: 2%;
+    
+    padding-left: 6%;
+    position: relative;
+    text-align: left;
+    justify-content: center;
+    font-weight: bold;
+    font-size: calc(20px + 1.0vw);
+}
+
+.body--content{
+    background-color: white;
+    font-size: calc(20px + 1.0vw);
+    margin-left: 10%;
+    margin-right: 10%;
+    margin-top: 10%;
+    margin-bottom: 10%;
+    text-align: center;
+
+}
