+++ neuro-app/src/pages/ViewMaterialsPages/ViewLessonMaterialsPage/ViewLesson.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import ViewLessonMaterials from '../../../components/LessonMaterials/ViewLessonMaterials';
+
+
+const ViewLesson: React.FC<{}> = () => {
+  
+
+  return (
+    <div>
+      <ViewLessonMaterials id_course="5" />
+    </div>
+  );
+};
+
+export default ViewLesson;
